include libgen directive
Despite that somehow in GDB the basename function seems to operate just fine without having the libgen.h header. There was a hard-to-find issue where the program would SIGSEGV in libc....randomly on bname assignage. After adding that to the file to be included seems to wholly fix the issue thus far in testing